+++ packages/cli/src/constants.ts
@@ -8,3 +8,5 @@ export const ROOT_API_URL_DEV = 'https://api.thechaindata.com';
 export const ROOT_API_URL_PROD = 'https://api.subquery.network';
 
 export const BASE_PROJECT_URL = 'https://project.subquery.network';
+
+export const BASE_TEMPLATE_URl = 'https://templates.subquery.network';