+++ packages/node-core/src/indexer/entities/index.ts
@@ -3,4 +3,3 @@
 
 export * from './Poi.entity';
 export * from './Metadata.entity';
-export * from './Mmr.entitiy';