+++ packages/types/src/interfaces.ts
@@ -16,9 +16,22 @@ export type FunctionPropertyNames<T> = {
   [K in keyof T]: T[K] extends Function ? K : never;
 }[keyof T];
 
+type SingleOperators = '=' | '!=';
+type ArrayOperators = 'in' | '!in';
+export type FieldOperators = SingleOperators | ArrayOperators;
+
+export type FieldsExpression<T> =
+  | [field: keyof T, operator: SingleOperators, value: T[keyof T]]
+  | [field: keyof T, operator: ArrayOperators, value: Array<T[keyof T]>];
+
 export interface Store {
   get(entity: string, id: string): Promise<Entity | undefined>;
   getByField(entity: string, field: string, value: any, options?: {offset?: number; limit?: number}): Promise<Entity[]>;
+  getByFields<T extends Entity>(
+    entity: string,
+    filter: FieldsExpression<T>[],
+    options?: {offset?: number; limit?: number}
+  ): Promise<T[]>;
   getOneByField(entity: string, field: string, value: any): Promise<Entity | undefined>;
   set(entity: string, id: string, data: Entity): Promise<void>;
   bulkCreate(entity: string, data: Entity[]): Promise<void>;