+++ packages/node-core/CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- Update apollo-links to 1.0.2, fix dictionary resolver failed to get token issue
 - Use test runs as unit for tests instead of entity checks (#1957)
 - handle APIs in connection pool whose initialization failed (#1970)
+- Fix generated operation hash with single entity, buffer did not get hashed issue.
+- Infinite recursion in setValueModel with arrays (#1993)
 
 ### Changed
 - Move more code from node to node-core. Including configure module, workers (#1797)
@@ -15,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Project upgrades feature and many other changes to support it (#1797)
-- skipBlock option to NodeConfig (#1968)
+- `skipBlock` option to NodeConfig (#1968)
+- `getByFields` to store (#1993)
 
 ## [4.2.3] - 2023-08-17
 ### Fixed